move gyrofrequency into plasmapy.formulary.frequencies
